smart_cane_backend/requirements.txt
flask
ultralytics
opencv-python
torch
pillow
lask-cors